Guardar transaccion
Guardar transaccion

+++ App.core.trans/App.core.trans/App.core.trans/Models/TransaccionTipoMovimiento.cs
@@ -11,6 +11,6 @@ namespace App.core.trans.Models
 		public string CodigoTipoMovimiento { get; set; }
 		public bool Estaactivo { get; set; }
 		public int Numeroverificador { get; set; }
-		public int ValueInsert { get; set; }
+		public decimal ValueInsert { get; set; }
 	}
 }
